requirements.txt
setuptools
Flask
Pillow
requests
opencv-python
mediapipe
numpy
rembg[cpu]
boto3